requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
email-validator==2.1.0
python-dotenv==1.0.1
bcrypt==4.1.2
pytest==8.0.0
httpx==0.26.0
google-generativeai==0.8.3
streamlit==1.31.0